move prototype decl and add a lot of tests
