+++ src/csim/update_ops_named_Z.c
@@ -17,11 +17,6 @@
 //void Z_gate_single(UINT target_qubit_index, CTYPE *state, ITYPE dim);
 //void Z_gate_parallel(UINT target_qubit_index, CTYPE *state, ITYPE dim);
 
-void Z_gate_single_unroll(UINT target_qubit_index, CTYPE *state, ITYPE dim);
-void Z_gate_single_simd(UINT target_qubit_index, CTYPE *state, ITYPE dim);
-void Z_gate_parallel_unroll(UINT target_qubit_index, CTYPE *state, ITYPE dim);
-void Z_gate_parallel_simd(UINT target_qubit_index, CTYPE *state, ITYPE dim);
-
 void Z_gate(UINT target_qubit_index, CTYPE *state, ITYPE dim) {
 	//Z_gate_old_single(target_qubit_index, state, dim);
 	//Z_gate_old_parallel(target_qubit_index, state, dim);
